+++ internal/providergen/data_sources/metric_monitor.ts
@@ -120,7 +120,7 @@ export default {
     },
     {
       name: "time_window_seconds",
-      type: "int",
+      type: "int64",
       description: "The time window in seconds to use for the aggregate query.",
       computedOptionalRequired: "computed",
       skipFill: true,
@@ -149,7 +149,7 @@ export default {
         },
         {
           name: "comparison_delta",
-          type: "int",
+          type: "int64",
           description:
             "The comparison delta in seconds to use for the aggregate query. Only available for `percent` type.",
           computedOptionalRequired: "computed",
@@ -212,7 +212,7 @@ export default {
             },
             {
               name: "condition_result",
-              type: "int",
+              type: "int64",
               description:
                 "When the condition is met, the result will be set to this value.",
               computedOptionalRequired: "computed",
